# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Auction",
        "description": "Documentation Auction web ",
        "contact": {
            "email": "rustamovvramziddin@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/v1/current-bid": {
            "get": {
                "tags": [
                    "Current-Bid"
                ],
                "summary": "Return only fields ",
                "description": "Get all data",
                "operationId": "3c66e6162e2031de4c9763a332f7822a",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Current-Bid"
                ],
                "summary": "Post a new data",
                "description": "Post new user data",
                "operationId": "f7b9d0b46eba8278de0471447dfd3d9f",
                "requestBody": {
                    "description": "Pass CurrentBid   credentials",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "product_id",
                                    "price"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "number",
                                        "format": "text",
                                        "example": "1"
                                    },
                                    "product_id": {
                                        "type": "number",
                                        "format": "",
                                        "example": "1"
                                    },
                                    "price": {
                                        "type": "text",
                                        "format": "text",
                                        "example": "$12000"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/v1/current-bid/{current-bid}": {
            "get": {
                "tags": [
                    "Current-Bid"
                ],
                "summary": "Get one ",
                "description": "Return all date related to ID{bid id}",
                "operationId": "05e6002d7464c6c3bdf3a6a8abaa88b4",
                "parameters": [
                    {
                        "name": "current-bid",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Current-Bid"
                ],
                "summary": "Get one and Delete related to Bid id",
                "description": "Return  date related to ID of the Bid",
                "operationId": "17bea12ac1b004a15b962d08989c2f72",
                "parameters": [
                    {
                        "name": "current-bid",
                        "in": "path",
                        "description": "put bid id and try to delete ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/v1/current-bid/{id}": {
            "put": {
                "tags": [
                    "Current-Bid"
                ],
                "summary": "Update current-bid ",
                "description": "Returns updated Bid",
                "operationId": "Update current-bid price of the product",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "bidding current update ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "description": "Update bid History",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "product_id",
                                    "price"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "number",
                                        "format": "text",
                                        "example": "5"
                                    },
                                    "product_id": {
                                        "type": "number",
                                        "format": "text",
                                        "example": "4"
                                    },
                                    "price": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "$4000"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Bid Updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "api": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": {
        "api": {
            "type": "http",
            "description": "Authorization token obtained from logging in.",
            "name": "Authorization",
            "in": "header",
            "scheme": "bearer"
        }
    }
}